Change to random access for read phase

+++ mmap-minor-faults.cc
@@ -4,10 +4,15 @@
 #include <sys/wait.h>
 #include <signal.h>
 
+#include <chrono>
 #include <iostream>
 #include <sstream>
+#include <string>
 
 #include <glog/logging.h>
+
+#include "uniform_random.hpp"
+#include "CLI11.hpp"
 
 constexpr uint64_t CACHELINE_SIZE = 64; 
 
@@ -18,10 +23,13 @@ void load_memory(char *data, uint64_t capacity) {
   }
 }
 
+foedus::assorted::UniformRandom rng(0xdeedbeef);
+
 void check_memory(char *data, uint64_t capacity) {
   for (uint64_t i = 0; i < capacity; i += CACHELINE_SIZE) {
-    char c = i;
-    CHECK_EQ(data[i], c);
+    size_t random_idx = CACHELINE_SIZE * rng.uniform_within(0, capacity / CACHELINE_SIZE);
+    char c = random_idx;
+    CHECK_EQ(data[random_idx], c);
   }
 }
 
@@ -38,7 +46,7 @@ void *mmap_memory(uint64_t capacity) {
 
 void *fine_grained_mmap_memory(uint64_t capacity) {
   auto data = reinterpret_cast<char *>(mmap(nullptr, capacity, PROT_NONE, flags, -1, 0));
-  auto step = 2 * 1024 * 1024;
+  auto step = 2 * 1024 * 1024;  // 2M
   for (uint64_t i = 0; i < capacity; i += step) {
     if (i + step > capacity) {
       auto ret = mprotect(data + i, capacity - i, PROT_READ | PROT_WRITE);
@@ -58,10 +66,18 @@ void *bulk_mmap_memory(uint64_t capacity) {
   return data;
 }
 
-int main() {
+int main(int argc, char **argv) {
+  CLI::App app{"mmap minor faults experiment"};
+
+  std::string profiling = "none";
+  app.add_option("--profiling", profiling, "What kind of profiling to run: none, stat, record");
+
+  CLI11_PARSE(app, argc, argv);
+
   uint64_t capacity = 16UL *
     1024 * 1024 * 1024;  // 1GiB
 
+    std::cout << "loading..." << std::endl;
 #ifdef MMAP
   auto mmap_data = reinterpret_cast<char *>(mmap_memory(capacity));
 #endif
@@ -88,19 +104,34 @@ int main() {
   load_memory(bulk_mmap_data, capacity);
 #endif
 
-  std::stringstream parent_pid;
-  parent_pid << getpid();
-
   pid_t perf_pid;
-  pid_t pid = fork();
-  if (pid == 0) {
-    exit(execl(
-          "/usr/bin/perf", "perf", "stat", "-B", "-e",
-          "cache-references,cache-misses,cycles,instructions,branches,faults,major-faults,minor-faults",
-          "-p", parent_pid.str().c_str(), nullptr));
+  if (profiling == "none") {
+    std::cout << "performance run..." << std::endl;
   } else {
-    perf_pid = pid;
+    std::stringstream parent_pid;
+    parent_pid << getpid();
+    pid_t pid = fork();
+
+    if (pid == 0) {
+      if (profiling == "stat") {
+        std::cout << "running perf stat..." << std::endl;
+        exit(execl(
+              "/usr/bin/perf", "perf", "stat", "-B", "-e",
+              "cache-references,cache-misses,cycles,instructions,branches,faults,major-faults,minor-faults",
+              "-p", parent_pid.str().c_str(), nullptr));
+      } else if (profiling == "record") {
+        std::cout << "running perf record..." << std::endl;
+        exit(execl(
+              "/usr/bin/perf", "perf", "record", "-F", "99", "-e",
+              "cache-references,cache-misses,cycles,instructions,branches,faults,major-faults,minor-faults",
+              "-p", parent_pid.str().c_str(), nullptr));
+      }
+    } else {
+      perf_pid = pid;
+    }
   }
+
+  auto starttime = std::chrono::system_clock::now();
 
   int repeat = 5;
   while (repeat-->0) {
@@ -122,7 +153,13 @@ int main() {
 #endif
   }
 
-  std::cerr << "stop perf..." << std::endl;
-  kill(perf_pid, SIGINT);
-  waitpid(perf_pid, nullptr, 0);
+  auto duration = std::chrono::duration<double>(std::chrono::system_clock::now() - starttime);
+
+  if (profiling == "stat" || profiling == "record") {
+    std::cerr << "stop perf..." << std::endl;
+    kill(perf_pid, SIGINT);
+    waitpid(perf_pid, nullptr, 0);
+  }
+
+  std::cout << "running time: " << duration << std::endl;
 }
